Configure target assemblies via attribute
Instead of defining namespaces in a config file, the generator now examines the assembly for the presence of the "PandorityTarget" attribute.

This approach is much safer and more performant than attempting to define a user config file. The big issue is, that the generate runs in different processes:
- Unity csc compiler (standard)
- Unity VBCS compiler (analyzers)
- Rider Roslyn worker

The working directory or executing assembly location is not consistent between these contexts, because e.g. Rider makes use of the "Shadow Copy" feature for assemblies. This leads to the Pandority.dll sometimes running in completely different temp directories, which is intended as such, but makes it very difficult to obtain a file path to the project.