Use absolute file paths for crash log
Using a relative file path is a bad idea because the context in which the DLL runs can be different. In Unity, the current directory will be the project, however, when e.g. Rider starts its own worker process to analyze code, the Pandority assembly will be copied to a temporary location ('shadow copying' feature). The same issue also affects API calls like Assembly.GetExecutingAssembly().Location, they all will point to different paths depending on the parent process.

After experimenting with different methods to attach a debugger (which is still very difficult), I came to the conclusion that a crash log is a reasonable development aid, even for the future.